Implement Dumpstate Board HAL for Trout
This HAL is for dumping the logs from PVM. It will dump the PVM logs into a
tar file and name it to "dumpstate_board.bin" under the bugreport zip
file.

Bug: 143787037
Test: build; create the directory /data/bugreport_test manually, and put
some files under it; `adb bugreport <zip_filename>` to generate the
report; rename the `dumpstate_board.bin` file to a tar file, the contents of the test
directory are packed here.

Change-Id: I82eab771c735c856291a5a3715aabf5b5085ab27