Implement Dumpstate Board HAL for Trout
This HAL is for dumping the logs from PVM. It will dump the PVM logs into a
tar file and name it to "dumpstate_board.bin" under the bugreport zip
file.

Bug: 143787037
Test: test with ag/10812717; build; create the directory /data/bugreport_test manually, and put
some files under it; `adb bugreport <zip_filename>` to generate the
report; rename the `dumpstate_board.bin` file to a tar file, the contents of the test
directory are packed here.

Change-Id: I8136ec1214cc378985f24cc3473829ad990822c9