Factor common code to read/write IIO attributes
A lot of the code to do with reading and writing IIO attributes is shared
across different data types, and can be templated away.

This still does not make for an entirely type-safe API as the format string
also needs to be derived from the value type, but this can be done in a
follow-up CL, as it's not a regression from the original code.

Bug: None
Test: flash and boot; KitchenSink
Change-Id: I7573f36a7ef52cec36385d9869b6c83941ed08e4