Edit Audio HAL to remove non-inclusive lingo
Android has guidelines at
https://source.android.com/setup/contribute/respectful-code
on how to write code that abides by the project's values of
equal dignity and inclusion for all.

Some of the terms in Audio HAL run counter to this goal.
Where feasible, use more appropriate terms.

Please note that some instances are API in tinyalsa and
others that we depend on and as such we can't change them
without breaking the API contract at this time.

Bug: 162964571
Test: build; Ctrl+F
Change-Id: I995939a9dd11e3bbcb5cbdf8f63ed4a876a6da20