Audio HAL: Adjust default buffer size.
The audio HAL works through tinyalsa, that doesn't provide way to query
supported parameters from the device. For this reason, all audio HALs in
Android use hardcoded audio parameters. In turn, default buffer size
that is allocated by the CF HAL does not match supported by the virtual
sound card. It leads to a lot of underruns and so on.

Bug: 167390806
Test: Build and boot
Change-Id: I6feda36b4e961539555e082e2e6b2d59c73d8f7f