Only include trout-kernel from the source tree when present
trout-kernel is not currently shipped in the generic PDK and is not
made available in AOSP. Taking a hard build dependency on it breaks
the trout userland build for partners who don't have our prebuilt
kernel binary.

The trout kernel can be built from source or packaged by means
that are hypervisor-specific.

Bug: 209814981
Test: build
Change-Id: If7596ca64f7da5225bae2611a588888addc32085