Set BOARD_BOOT_HEADER_VERSION before including other makefiles
BOARD_MKBOOTIMG_ARGS was appended to with +=, which was
defaulting to using deferred expansion instead of simple
expansion because it was never assigned to before. The
deferred expansion was then used in certain products
to change BOARD_BOOT_HEADER_VERSION later on and have
the change be reflected in BOARD_MKBOOTIMG_ARGS.

Deferred expansion will no longer work after switching
to starlark, so instead have products set BOARD_BOOT_HEADER_VERSION
before including this makefile, and change this makefile
to not override the BOARD_BOOT_HEADER_VERSION that was
set elsewhere.

Bug: 201700692
Test: Verified that this doesn't cause any changes to the ninja files on the affected products.
Change-Id: I929c7d22ff492c3e974acb3a33832c3c1966b580