Select kernel 5.10 for trout
If no kernel version is selected, the build rules fail to find the right
kernel binary for trout and default to the prebuilt kernel for cuttlefish
which is not able to boot trout to userspace

The easiest fix is to specify a preferred kernel version, which lets the
wildcard check complete, and produce the right binary file

I do not claim sufficient expertise in the intricacies of soong and make
to produce a thorough diagnosis of the root cause of this issue, as other
rules work correctly with a similar defaulting, but this one clearly
does not. This patch fixes it.

Bug: 188702624
Test: build, look for kernel version string in boot.img
Change-Id: Icfd4308b7ca14a9811a06b58ba8272dc43113eee