Allow customizable EVS configuration
This commit fixes a regression from Android 12. The older version
of the code allowed the implementor to customize the EVS configuration
files, but this was lost during Android 13 development.

Restore the LOCAL_EVS_PRODUCT_COPY_FILES variable as a staging ground for
what files need to be copied, instead of hardcoding the default version

Bug: 251227885
Test: build and boot device
Change-Id: Ief3908e9953faa1da36f514f162ba11481dddcda